+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- `create` / `with` will no longer throw an exception if given an invalid config; rather, the config's invalid values
+  will be replaced with sensible defaults
 - Remove `atomic-primops` dependency
 
 ## [0.4.0.1] - 2022-11-05